@@ -1,17 +1,21 @@
 module github.com/lemonade-command/lemonade
 
-go 1.16
+go 1.22
 
 require (
-	github.com/BurntSushi/toml v0.3.1 // indirect
 	github.com/atotto/clipboard v0.1.4
-	github.com/go-stack/stack v1.8.0 // indirect
 	github.com/inconshreveable/log15 v0.0.0-20201112154412-8562bdadbbac
-	github.com/mattn/go-colorable v0.1.8 // indirect
-	github.com/mattn/go-isatty v0.0.13 // indirect
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/monochromegane/conflag v0.0.0-20151130130520-6d68c9aa4183
 	github.com/pocke/go-iprange v0.0.0-20150823054938-08fbe355c365
 	github.com/skratchdot/open-golang v0.0.0-20200116055534-eef842397966
+)
+
+require (
+	github.com/BurntSushi/toml v0.3.1 // indirect
+	github.com/go-stack/stack v1.8.0 // indirect
+	github.com/mattn/go-colorable v0.1.8 // indirect
+	github.com/mattn/go-isatty v0.0.13 // indirect
+	golang.org/x/sys v0.0.0-20200223170610-d5e6a3e2c0ae // indirect
 	gopkg.in/yaml.v2 v2.4.0 // indirect
 )
